refactor(travel-card): group skeleton components under skeleton/
