+++ src/screen-components/travel-card/index.ts
@@ -1,7 +1,9 @@
 export {TravelCard} from './TravelCard';
-export {TravelCardSkeleton} from './TravelCardSkeleton';
-export {WithTravelCardSkeleton} from './WithTravelCardSkeleton';
-export {SkeletonBlock} from './SkeletonBlock';
+export {
+  SkeletonBlock,
+  TravelCardSkeleton,
+  WithTravelCardSkeleton,
+} from './skeleton';
 export {getTripPatternStatus} from './utils';
 export {type TravelCardHeaderSize} from './TravelCardHeader';
 import {withCompositeAccessibility} from '@atb/modules/composite-accessibility';
