+++ src/screen-components/travel-card/legs/index.ts
@@ -1,3 +1,2 @@
 export {TransportationLeg} from './TransportationLeg';
 export {FootLeg} from './FootLeg';
-export {LegsSkeleton} from './LegsSkeleton';
