+++ src/screen-components/travel-card/skeleton/SkeletonBlock.tsx
