+++ src/screen-components/travel-card/skeleton/index.ts
@@ -1,0 +1,3 @@
+export {SkeletonBlock} from './SkeletonBlock';
+export {TravelCardSkeleton} from './TravelCardSkeleton';
+export {WithTravelCardSkeleton} from './WithTravelCardSkeleton';
